algo

+++ inode.c
@@ -30,40 +30,40 @@ int inode_iget(struct unixfilesystem *fs, int inumber, struct inode *inp) {
  */
 int inode_indexlookup(struct unixfilesystem *fs, struct inode *inp, int blockNum) {  
     //Implement code here
-    int inode_file_size = inode_getsize(inp);
+    uint16_t inode_file_size = inode_getsize(inp);
     if (inode_file_size == 0) return -1;
-    int inode_blocks = inode_file_size/DISKIMG_SECTOR_SIZE;
+    uint16_t inode_blocks = inode_file_size/DISKIMG_SECTOR_SIZE;
     if (blockNum < 0 || blockNum >= inode_blocks) return -1;
     if (inode_file_size <= DISKIMG_SECTOR_SIZE*8){
         return inp->i_addr[blockNum];
     }
-    int indir_block_num = blockNum / 256;
-    int offset = blockNum % 256;
-    int *indir_block = malloc(DISKIMG_SECTOR_SIZE);
+    uint16_t indir_block_num = blockNum / 256;
+    uint16_t offset = blockNum % 256;
+    uint16_t *indir_block = malloc(DISKIMG_SECTOR_SIZE);
     if (indir_block_num < 7){
         if (diskimg_readsector(fs->dfd, inp->i_addr[indir_block_num], indir_block) == -1) {
         free(indir_block);
         return -1;
         } else {
-            int block = indir_block[offset];
+            uint16_t block = indir_block[offset];
             free(indir_block);
             return block;
         }
     }
     if (indir_block_num >= 7){
-        int indir_block_num = (blockNum - 7*256) / 256;
-        int offset = (blockNum - 7*256) % 256;
+        uint16_t indir_block_num = (blockNum - 7*256) / 256;
+        uint16_t offset = (blockNum - 7*256) % 256;
         if (diskimg_readsector(fs->dfd, inp->i_addr[7], indir_block) == -1) {
             free(indir_block);
         return -1;
         } else {
-            int *second_indir_block = malloc(DISKIMG_SECTOR_SIZE);
+            uint16_t *second_indir_block = malloc(DISKIMG_SECTOR_SIZE);
             if (diskimg_readsector(fs->dfd, indir_block[indir_block_num], second_indir_block) == -1) {
                 free(second_indir_block);
                 free(indir_block);
                 return -1;
             }
-            int block = second_indir_block[offset];
+            uint16_t block = second_indir_block[offset];
             free(indir_block);
             free(second_indir_block);
             return block;
